🚨 Fix(#197): claer 페이지 -> logout 페이지
